Move calldata code to the new module
