+++ solidity-verifiers/src/utils/mod.rs
@@ -1,32 +1,8 @@
 use crate::{GPL3_SDPX_IDENTIFIER, PRAGMA_GROTH16_VERIFIER};
 use askama::Template;
-use crypto::{digest::Digest, sha3::Sha3};
-use num_bigint::BigUint;
+
 pub mod calldata;
 pub mod encoding;
-
-/// Formats call data from a vec of bytes to a hashmap
-/// Useful for debugging directly on the EVM
-/// !! Should follow the contract's function signature, we assuming the order of arguments is correct
-pub fn get_formatted_calldata(calldata: Vec<u8>) -> Vec<String> {
-    let mut formatted_calldata = vec![];
-    for i in (4..calldata.len()).step_by(32) {
-        let val = BigUint::from_bytes_be(&calldata[i..i + 32]);
-        formatted_calldata.push(format!("{}", val));
-    }
-    formatted_calldata
-}
-
-/// Computes the function selector for the nova cyclefold verifier.
-/// It is computed on the fly since it depends on the IVC state length.
-pub fn get_function_selector_for_nova_cyclefold_verifier(state_len: usize) -> [u8; 4] {
-    let mut hasher = Sha3::keccak256();
-    let fn_sig = format!("verifyNovaProof(uint256[{}],uint256[4],uint256[2],uint256[3],uint256[2],uint256[2][2],uint256[2],uint256[4],uint256[2][2])", state_len * 2 + 1);
-    hasher.input_str(&fn_sig);
-    let hash = &mut [0u8; 32];
-    hasher.result(hash);
-    [hash[0], hash[1], hash[2], hash[3]]
-}
 
 #[derive(Template)]
 #[template(path = "header_template.askama.sol", ext = "sol")]
